ARROW-12425: [Rust] Fix new_null_array dictionary creation
It is my understanding that an arrow array should always have a backing values array, even if the content is all nulls. new_null_array currently violates this as it doesn't allocate the backing store for DictionaryArrays. This causes the concat kernel, and possibly others, to panic with index violations

Signed-off-by: Raphael Taylor-Davies <r.taylordavies@googlemail.com>

Closes #10072 from tustvold/null-dictionary-creation

Authored-by: Raphael Taylor-Davies <r.taylordavies@googlemail.com>
Signed-off-by: Andrew Lamb <andrew@nerdnetworks.org>