ARROW-12598: [C++][Dataset] Speed up CountRows for CSV
This does not implement a fast path for CSV. However, it does configure the CSV reader to not actually deserialize any data, resulting in a large gain. When scanning 85 million rows of the NYC Taxi dataset, scan time dropped from ~7 seconds to 2.

This also sneaks in an implementation of the fast path for InMemoryFragment.

Closes #10270 from lidavidm/arrow-12598

Authored-by: David Li <li.davidm96@gmail.com>
Signed-off-by: Antoine Pitrou <antoine@python.org>