ARROW-12007: [C++] Loading parquet file returns "Invalid UTF8 payload" error
Judging from the comment "avoid spending time validating UTF8 data" with the setting of the false value to the cast_options.allow_invalid_utf8, it seems to me this was intended to be true rather than false.

Also, this resolved the error I was getting through the arrow R package, which seems to be ARROW-12007.

Closes #10759 from hideaki/cancel_unnecessary_utf8_check

Authored-by: Hideaki Hayashi <hihayash@gmail.com>
Signed-off-by: Antoine Pitrou <antoine@python.org>