MINOR: Rename setup options chunk
There is a typo in the "Arrow R Developer Guide" where there is a space in the name of a code chunk instead of a hyphen.  I'm doing some automated analyses of the different code chunks and can't parse this one due to the typo.  This PR just adds a hyphen.

Closes #10754 from thisisnic/rmd_label

Authored-by: Nic Crane <thisisnic@gmail.com>
Signed-off-by: Jonathan Keane <jkeane@gmail.com>