ARROW-12072: [Go] Fix panics in ipc writer for sliced records
Figuring out the length of the truncated values array and offsets was being calculated incorrectly by mixing the length of elements and raw byte length and not utilizing the data offset when creating the slice for shifting the zero-based offsets. This fixes both issues so that writing sliced records of FixedWidthDataType arrays and arrays with offsets now works correctly.

Closes #11421 from zeroshade/arrow-12072-ipc-slicing

Authored-by: Matthew Topol <mtopol@factset.com>
Signed-off-by: Matthew Topol <mtopol@factset.com>