ARROW-14239: [R] Don't use rlang::as_label
https://github.com/r-lib/rlang/pull/1277 changed the behavior of `as_label` when the call contains operators. On current release of rlang, we see:

```
> as_label(as_quosure(~nchar(chr, type = "bytes", allowNA = TRUE) == 1))
[1] "nchar(chr, type = \"bytes\", allowNA = TRUE) == 1"
```

On dev rlang, we get

```
> as_label(as_quosure(~nchar(chr, type = "bytes", allowNA = TRUE) == 1))
[1] "... == 1"
```

This causes a test failure, which we were told that we needed to address in advance of their release.

According to the tests added in that rlang PR, this is the intended behavior of `as_label` now, but it doesn't seem to be what we want. We can recover the behavior we're expecting by using `deparse()`.

Closes #11338 from nealrichardson/drop-as-label

Authored-by: Neal Richardson <neal.p.richardson@gmail.com>
Signed-off-by: Jonathan Keane <jkeane@gmail.com>