ARROW-14654: [R][Docs] Add article on how to run R with C++ debugger to dev docs
This PR adds an article about how to run R with the C++ debugger to the dev docs.  It also adds a submenu for articles to be published on the `pkgdown` site but not distributed with the package as vignettes.  Note that the current R developer guide remains as it's own menu option so we don't break links to it in other packages.

Closes #11658 from thisisnic/ARROW-14654_debugger_docs

Authored-by: Nic Crane <thisisnic@gmail.com>
Signed-off-by: Nic Crane <thisisnic@gmail.com>