ARROW-14022: [Dev] Remove arrow/dev/benchmarking
The following directory contains a WIP that hasn't been modified in 3 years [1]:

- arrow/dev/benchmarking

The directory with benchmarking code that is currently in use is here:

- arrow/dev/archery/archery/benchmark

I was asked to answer questions about Arrow benchmarking today, and I naturally traversed into the higher level, WIP, benchmarking directory, but the one in use is nested deeper. Perhaps it's time to remove the WIP directory.

PS. This WIP was the jumping off point for Conbench (https://github.com/conbench/conbench). Much appreciated!

[1] Sadly, the author passed away. Thank you Tanya! RIP ;(

https://www.alumni.caltech.edu/in-memoriam/2019-9-24-tanya-schlusser-bs-99-1

Closes #11173 from dianaclarke/ARROW-14022

Authored-by: Diana Clarke <diana.joan.clarke@gmail.com>
Signed-off-by: Neal Richardson <neal.p.richardson@gmail.com>