ARROW-14840: [R][CI] test-ubuntu-20.10-docs nightly build failing due to R install issue
I'm not 100% sure why this fixes this fail, but downgrading the Ubuntu version to 20.04 seems to fix it, and here's my reasoning.

So, apt-key, which is used as part of the R bit of the docs build is now deprecated[1] and Ubuntu 20.10 (Groovy Gorilla) is no longer a supported release as of 22nd July 2021 [2] which can apparently cause issues subsequently trying to do stuff via apt[3].

Reading the R docs for Ubuntu packages[4], it says there that "Package for the current R 4.1. release* are available for most stable Desktop releases of Ubuntu until their official end of life date" which maybe explains why the URL for R packages works fine for focal[5] but results in a 404 error for groovy [6].

The way of getting the signing key for updating packages changed [4] to no longer use apt-key [7] - I have no idea if this had any effect either, but to be on the safe side I have updated that too.

[1] https://manpages.debian.org/testing/apt/apt-key.8.en.html
[2] https://wiki.ubuntu.com/Releases
[3] https://news.itsfoss.com/ubuntu-20-10-end-of-life/
[4] https://cran.r-project.org/bin/linux/ubuntu/
[5] https://cran.r-project.org/bin/linux/ubuntu/focal-cran40/
[6] https://cran.r-project.org/bin/linux/ubuntu/groovy-cran40/
[7] http://web.archive.org/web/20210618164400/https://cloud.r-project.org/bin/linux/ubuntu/

Closes #11815 from thisisnic/ARROW-14840

Authored-by: Nic Crane <thisisnic@gmail.com>
Signed-off-by: Jonathan Keane <jkeane@gmail.com>