ARROW-14936: [C++][Gandiva] Fix split_part function in gandiva
Split_part function sporadically returns error. This happens when the output string is empty. The call to gdv_fn_context_arena_malloc with size 0 returns nullptr when there has been no earlier call which leads to the error.
The allocation here is also redundant since we can return slice of the input string

Closes #11819 from projjal/splitpart

Authored-by: Projjal Chanda <iam@pchanda.com>
Signed-off-by: Pindikura Ravindra <ravindra@dremio.com>