ARROW-14482: [C++][Gandiva] Implement MASK_FIRST_N and MASK_LAST_N functions
**MASK_FIRST_N**

Returns a masked version of str with the first n values masked. Upper case letters are converted to "X", lower case letters are converted to "x" and numbers are converted to "n". For example, mask_first_n("1234-5678-8765-4321", 4) results in nnnn-5678-8765-4321.

**MASK_LAST_N**

Returns a masked version of str with the last n values masked. Upper case letters are converted to "X", lower case letters are converted to "x" and numbers are converted to "n". For example, mask_last_n("1234-5678-8765-4321", 4) results in 1234-5678-8765-nnnn.

Closes #11551 from augustoasilva/feature/add-mask-first-n-and-mask-last-n

Authored-by: Augusto Silva <augusto.a.silva@hotmail.com>
Signed-off-by: Pindikura Ravindra <ravindra@dremio.com>